fix: improve import-file-parse error handling
