+++ .vscode/settings.json
@@ -2,9 +2,11 @@
   "cSpell.words": [
     "BACKORDER",
     "Backorders",
+    "microservices",
     "middyfy",
     "Millis",
     "monorep",
+    "Monorepository",
     "querystrings",
     "Tsjson"
   ],
