+++ import-service/package.json
@@ -1,6 +1,6 @@
 {
   "name": "import-service",
-  "version": "1.6.3",
+  "version": "2.0.3",
   "description": "API-based microservice for importing products",
   "main": "serverless.ts",
   "scripts": {
