+++ import-service/src/functions/import-file-parser/handler.ts
@@ -10,23 +10,27 @@ import { importFileParser, TImportProcessResult } from './import-file-parser';
 const handler: Handler<S3Event, APIGatewayProxyResult> = async (event) => {
   DefaultLogger.trace(event, 'import-file-parser');
 
-  let result: TImportProcessResult;
+  let result: TImportProcessResult = { isSuccessful: false };
 
   try {
     result = await importFileParser(event);
   } catch (error) {
-    DefaultLogger.error('Lambda handler exception');
+    DefaultLogger.error('ImportFileParser', error);
 
     switch (error) {
       default:
-        return formatJSONResponse({
-          ...error,
-          message: error.message ?? 'failed',
-        });
+        return formatJSONResponse(
+          {
+            ...result,
+            ...error,
+            message: error.message ?? 'failed',
+          },
+          202,
+        );
     }
   }
 
-  return formatJSONResponse({ message: 'success' });
+  return formatJSONResponse({ ...result }, result.isSuccessful ? 200 : 202);
 };
 
 export const main = middyfy(handler);
