+++ import-service/src/model/index.ts
@@ -4,4 +4,7 @@ export enum ErrorMessages {
   EmptyRecordList = 'Empty record list',
   SomethingWentWrong = 'Something went wrong',
   ParsedDataIsInvalid = 'Parsed data is invalid',
+  SQSUrlWasNotProvided = 'SQS URL was not provided',
+  SendMessageFailed = 'Send message failed',
+  GetObjectFailed = 'Get object failed',
 }
